Add a median filter C implementation

+++ utils.inline.c
@@ -2,7 +2,7 @@
 
 #ifdef _WIN32
     #include <windows.h>
-#elif MACOS
+#elif __APPLE__
     #include <sys/param.h>
     #include <sys/sysctl.h>
 #else
@@ -20,8 +20,8 @@ static size_t utils_get_number_of_cpu_cores()
     SYSTEM_INFO info;
     GetSystemInfo(&info);
 
-    result = info.dwNumberOfProcessors;
-#elif MACOS
+    result = (int) info.dwNumberOfProcessors;
+#elif __APPLE__
     int nm[2];
 
     size_t length = 4;
@@ -37,9 +37,9 @@ static size_t utils_get_number_of_cpu_cores()
         sysctl(nm, 2, &count, &length, NULL, 0);
     }
 
-    result = count;
+    result = (int) count;
 #else
-    result = sysconf(_SC_NPROCESSORS_ONLN);
+    result = (int) sysconf(_SC_NPROCESSORS_ONLN);
 #endif
 
     if (result < 1) {
